requirements.txt
# 基础依赖  
fastapi==0.104.1  
uvicorn==0.23.2  
pydantic==2.4.2  
python-multipart==0.0.6  
numpy==1.24.3  
pandas==2.0.3  
requests==2.31.0  

# PDF处理  
PyMuPDF==1.23.5  

# 向量数据库  
pymilvus==2.3.0  

# 图数据库  
neo4j==5.13.0  

# 深度学习  
torch==2.0.1  
transformers==4.34.0  
pytorch-crf==0.7.2  
scikit-learn==1.3.2  

# 中文处理  
jieba==0.42.1  

# API请求  
zhipuai==1.0.7  

# 工具和调试  
tqdm==4.66.1  
matplotlib==3.7.3  
loguru==0.7.2  